Integrate prices test into new update product suite relying on unified command
